FarmX is an organisation aiming to raise awareness about the inefficiency and plights that farmers face today.
Our primary ambition is to reach as many as individuals as individuals possible in order to educate people about the perils of being a farmer today and means through which these problems can be remedied. We intend to collate and scrutinize perspectives from the myriad stakeholders of this locus in order to augment our understanding of the topic as well as the validity of our initiative.